Deep dive problem 4: Retrieve Weather.

+++ hmb-functional-deep-dive/src/common/UserInput.scala
@@ -44,7 +44,7 @@ object UserInput {
   }
 
   // Prompt user for Double input.
-  def promptNonNegativeDouble(): Int = {
+  def promptNonNegativeDouble(): Double = {
     try {
       val userInput = scala.io.StdIn.readDouble()
       if (userInput >= 0) {
